#incomes-outcomes: create outcomes services

+++ src/main/resources/db/migration/V0.2__create_outcomes_table.sql
@@ -4,7 +4,7 @@ CREATE TABLE  outcomes (
   id uuid DEFAULT uuid_generate_v4(),
   user_id uuid NOT NULL,
   description VARCHAR NOT NULL,
-  value BIGINT NOT NULL,
+  value NUMERIC(19,2) NOT NULL,
   outcome_date DATE NOT NULL,
   updated_at TIMESTAMP NOT NULL DEFAULT NOW(),
